requirements.txt
streamlit==1.31.0
streamlit-mermaid==0.0.4
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==4.1.1
supabase==2.3.4
Flask==2.3.3

